+++ src/ColorScheme/ColorScheme.mjs
@@ -1,5 +1,0 @@
-/** @typedef {import("../Localization/Localization.mjs").Localization} Localization */
-
-/**
- * @typedef {{label: (localization: Localization) => Promise<string>, name: string}} ColorScheme
- */