+++ src/ColorScheme/ColorSchemeWithSystemColorScheme.mjs
@@ -1,6 +1,0 @@
-/** @typedef {import("./ColorScheme.mjs").ColorScheme} ColorScheme */
-/** @typedef {import("./SystemColorScheme.mjs").SystemColorScheme} SystemColorScheme */
-
-/**
- * @typedef {ColorScheme & {"system-color-scheme": SystemColorScheme | null}} ColorSchemeWithSystemColorScheme
- */