+++ src/ColorSchemeObject.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./Localization/Localization.mjs").Localization} Localization */
+
+/**
+ * @typedef {{label: (localization: Localization) => Promise<string>, name: string}} ColorSchemeObject
+ */