+++ src/ColorSchemeWithSystemColorScheme.mjs
@@ -1,0 +1,6 @@
+/** @typedef {import("./ColorSchemeObject.mjs").ColorSchemeObject} ColorSchemeObject */
+/** @typedef {import("./SystemColorScheme.mjs").SystemColorScheme} SystemColorScheme */
+
+/**
+ * @typedef {ColorSchemeObject & {"system-color-scheme": SystemColorScheme | null}} ColorSchemeWithSystemColorScheme
+ */