+++ src/SystemColorScheme.mjs
@@ -1,4 +1,4 @@
-/** @typedef {import("../Localization/Localization.mjs").Localization} Localization */
+/** @typedef {import("./Localization/Localization.mjs").Localization} Localization */
 
 /**
  * @typedef {{"default-color-scheme": string, detector: MediaQueryList, label: (localization: Localization) => Promise<string>, name: string, "use-in-color-scheme": boolean}} SystemColorScheme